==> Lessons/14_TimeSeries.Rmd <==
<<<<<<< HEAD

=======
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10
---
title: "14: Time Series Analysis"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2019"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Describe the aspects of hierarchical models, fixed effects, and random effects
2. Choose and justify appropriate statistical models when time is an explanatory variable
3. Apply Mann-Kendall and Seasonal Mann-Kendall to datasets with temporal components

## SET UP YOUR DATA ANALYSIS SESSION

```{r, message = FALSE, warning = FALSE}
getwd()
setwd("/Users/rachelbash/Documents/DUKE/Data Analytics/Environmental_Data_Analytics")
library(tidyverse)
install.packages("trend")
library(trend)


PeterPaul.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaul_Processed.csv")
USGS.flow.data <- read.csv("/Data/Raw/USGS_Site02085000_Flow_Raw.csv")

# Rename columns
colnames(USGS.flow.data) <- c("agency_cd", "site_no", "datetime", 
                              "discharge.max", "discharge.max.approval", 
                              "discharge.min", "discharge.min.approval", 
                              "discharge.mean", "discharge.mean.approval", 
                              "gage.height.max", "gage.height.max.approval", 
                              "gage.height.min", "gage.height.min.approval", 
                              "gage.height.mean", "gage.height.mean.approval")

# Set date to date format
PeterPaul.nutrients$sampledate <- as.Date(PeterPaul.nutrients$sampledate, 
                                               format = "%Y-%m-%d")
USGS.flow.data$datetime <- as.Date(USGS.flow.data$datetime, 
                              format = "%m/%d/%y")

mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)

```

## NONPARAMETRIC TREND TESTS

In many environmental datasets (especially climate and hydrology), we might not expect a linear trend in the response variable over time. In this case, we will need to employ a nonparametric test to determine whether there is a monotonic trend (i.e., consistent increase or decrease but not necessarily linear) over time. We will illustrate a few examples of nonparametric trend tests today with the `trend` package. 

A vignette for the `trend` package can be found here: https://cran.r-project.org/web/packages/trend/vignettes/trend.pdf. More details here: https://cran.r-project.org/web/packages/trend/trend.pdf.

We will run a Mann-Kendall and a Seasonal Mann-Kendall test today, but there are additional variants of these tests within the package including a correlated Seasonal Mann-Kendall test, a multivariate Mann-Kendall test, a partial Mann-Kendall test, a partial correlation trend test, and a Cox and Stuart trend test. Look into the documentation for these tests to determine which one is appropriate for your purposes.

### Mann-Kendall Test
<<<<<<< HEAD
A Mann-Kendall test will analyze whether there is a monotonic trend in the response variable over time. Let's use the Mann-Kendall test to investigate whether there is a trend in total phosphorus concentrations in Peter Lake over time. 

```{r}
# Wrangle our dataset
PeterPaul.nutrients.surface <- 
  PeterPaul.nutrients %>%
  select(-lakeid, -depth_id, -comments) %>%
  filter(depth == 0) %>%
  filter(!is.na(tp_ug))
```
A Mann-Kendall test will analyze whether there is a monotonic trend in the response variable over time. Let's use the Mann-Kendall test to investigate whether there is a trend in total phosphorus concentrations in Peter Lake over time. Mann-Kendall test does not account for seasonality.

  
```{r}
# Wrangle our dataset; looking at trends at top of the lake
PeterPaul.nutrients.surface.tp <- 
  PeterPaul.nutrients %>%
  select(-lakeid, -depth_id, -comments) %>% #take out lakeid, depth_id, and comment columns
  filter(depth == 0) %>% #depth is equal to 0, so looking at surface concentrations
  filter(!is.na(tp_ug)) #timeseries analysis runs into errors with NAs, so we can tell R to remove them
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10

# Initial visualization of data
ggplot(PeterPaul.nutrients.surface, aes(x = sampledate, y = tp_ug, color = lakename)) + 
  geom_point() +
  scale_color_manual(values = c("#7fcdbb", "#253494"))

# Split dataset by lake
Peter.nutrients.surface <- filter(PeterPaul.nutrients.surface, lakename == "Peter Lake")
Paul.nutrients.surface <- filter(PeterPaul.nutrients.surface, lakename == "Paul Lake")

# Run a Mann-Kendall test
mk.test(Peter.nutrients.surface$tp_ug)
<<<<<<< HEAD
=======
#Mann-Kendall test results: p-value is very low, so reject the null. Null in this test is that there is no trend over time. Z-score -- further away from 0 the z-score is, the direction in magnitude of the trend. Z is 4.3 here, so it's a positive direction. 
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10

```

However, it looks like there might be a breakpoint in our dataset. Further, we know that Peter Lake underwent experimental fertilization starting in May 1993, a perturbation which we might expect to have induced a regime shift in the ecosystem. In this case, we might want to find out whether there is a breakpoint, or changepoint, in our dataset. 

### Pettitt's Test
Pettitt's test is also included in the `trend` package. This nonparametric test will determine whether there is a shift in the central tendency of the time series and will tell us at what point the changepoint occurs (if it detects one). Note: Pettitt's Test will only test for one changepoint, and further tests must be run if multiple change points are suspected.
```{r}
# Test for change point
pettitt.test(Peter.nutrients.surface$tp_ug)
<<<<<<< HEAD

# Run separate Mann-Kendall for each change point
mk.test(Peter.nutrients.surface$tp_ug[1:34])
mk.test(Peter.nutrients.surface$tp_ug[35:132])

# Is there a second change point?
pettitt.test(Peter.nutrients.surface$tp_ug[35:132])

# Run another Mann-Kendall for the second change point
mk.test(Peter.nutrients.surface$tp_ug[35:113])
mk.test(Peter.nutrients.surface$tp_ug[114:132])

# Run the same test for Paul Lake. 
mk.test(Paul.nutrients.surface$tp_ug)
pettitt.test(Paul.nutrients.surface$tp_ug)

# Add vertical lines to the original graph to represent change points
ggplot(PeterPaul.nutrients.surface, aes(x = sampledate, y = tp_ug, color = lakename)) + 
  geom_point() +
  scale_color_manual(values = c("#7fcdbb", "#253494"))
=======
#results give us a p-value (> than 0.05, test has not detected a significant change point) and a probable change point; will have to look at where the change point happened in the dataset; pettitt test says change point is 35, which corresponds to row 35

# Run separate Mann-Kendall for each change point based on the change point value given in the test above
mk.test(Peter.nutrients.surface$tp_ug[1:34]) #p-value is close to 1 and z score is very low, so no trend detected between 1991 - 1993
mk.test(Peter.nutrients.surface$tp_ug[35:132]) #p-value is smaller than one above, and z-score is negative; trend is not significant because of the p-value

# Is there a second change point?
pettitt.test(Peter.nutrients.surface$tp_ug[35:132]) #there is a significant change point bc of low p-value; probable change point at 79; have to do 35+79 = 114, where 114 is the row number

# Run another Mann-Kendall for the second change point
mk.test(Peter.nutrients.surface$tp_ug[35:113])#trend from 1993 - 1997 and there is a significant positive trend because z-score is positive
mk.test(Peter.nutrients.surface$tp_ug[114:132])#trend from 1998 - 1999 and there is not a significant trend. 

# Run the same test for Paul Lake. 
mk.test(Paul.nutrients.surface$tp_ug) #adds up the difference of measurements 
pettitt.test(Paul.nutrients.surface$tp_ug) #no significant change point for this lake because the p-value is 0.1244

# Add vertical lines to the original graph to represent change points
ggplot(PeterPaul.nutrients.surface, aes(x = sampledate, y = tp_ug, color = lakename)) +
  geom_point() +
  scale_color_manual(values = c("#7fcdbb", "#253494")) +
  geom_vline(xintercept = as.Date("1993-05-26"), color="#253494", lty = 2) +
  geom_vline(xintercept = as.Date("1998-01-01"), color="#253494", lty = 2)

#can look at sens slope to figure out the slope of the line in the graphs
#key assumption we violated with this test is that mann-kendall thinks all points are independent of each other
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10

```
### Seasonal Mann-Kendall

<<<<<<< HEAD
Like a **Mann-Kendall Test**, the **Seasonal Mann-Kendall Test**, or **Hirsch-Slack Test**, analyzes trends in response variables over time. It replaces the traditional Mann-Kendall when there are seasonal trends in a dataset that obscure the overall direction of the trend. It is important to note that "seasonal" does not necessarily equate to actual seasons but can represent any time period within which there are oscillating temporal trends. The test needs at least two seasons to operate.
=======
Like a **Mann-Kendall Test**, the **Seasonal Mann-Kendall Test**, or **Hirsch-Slack Test**, analyzes trends in response variables over time. It replaces the traditional Mann-Kendall when there are seasonal trends in a dataset that obscure the overall direction of the trend. It is important to note that "seasonal" does not necessarily equate to actual seasons but can represent any time period within which there are oscillating temporal trends. The test needs at least two seasons to operate. Want two distinct seasons with seasonal mann-kendall test.
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10

For instance, we might want to know whether there is a change in discharge of the Eno River over the last 10 years.
```{r}
# Wrangle the USGS dataset
USGS.flow.data.trimmed <- USGS.flow.data %>%
  select(datetime, discharge.mean) %>%
  filter(datetime > as.Date("2008-12-31") & datetime < as.Date("2019-01-01"))

# Visualize the data
ggplot(USGS.flow.data.trimmed, aes(x = datetime, y = discharge.mean)) +
<<<<<<< HEAD
  geom_point(size = 0.5, alpha = 0.5)
=======
  geom_point(size = 0.5, alpha = 0.5) + scale_y_log10() #+
  #geom_abline(slope = Inf, intercept = as.Date("2010-01-01"))

>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10
```

### Interpolation
Some situations may require us to predict values for data points that fall within the time frame of our analyses but were not sampled. For instance, the `smk.test` function needs to take a time series format rather than a data frame, which cannot have any NAs. In this case, we will want to make an estimate of the missing values based on what we know about the dataset using a method called **interpolation.** There are several options for interpolation: 

<<<<<<< HEAD
  + **Means interpolation:** Defines values between sampled values as the mean value within a dataset. Uses the R function `aggregate`.
  
  + **Piecewise constant interpolation:** Defines values between sampled values as the value of the nearest sampled value. Uses the R function `approx` with `method = "constant"`
  
  + **Linear interpolation:** Defines values between sampled values based on the slope between sampled values. Uses the R function `approx` with `method = "linear"`
  
  + **Spline interpolation:** Defines values between sampled values based on polynomial functions between sampled values and chooses the polynomials so that they fit smoothly together. Uses the R function `splinefun`.

Question: Under what circumstances would you consider each of these options for interpolation? 

> ANSWER: 
=======
Seasonal Mann-Kendall test does not deal with NAs. So we might need to interpolate our data with interpolation methods above.

  + **Means interpolation:** Defines values between sampled values as the mean value within a dataset. Uses the R function `aggregate`. (i.e. mean humidity)
  
  + **Piecewise constant interpolation:** Defines values between sampled values as the value of the nearest sampled value. Uses the R function `approx` with `method = "constant"`. (also known as nearest neighbor interpolation. (i.e. going to find a place close to nearest neighbor)
  
  + **Linear interpolation:** Defines values between sampled values based on the slope between sampled values. Uses the R function `approx` with `method = "linear"`. (i.e. connect dots and predict based on the point)
  
  + **Spline interpolation:** Defines values between sampled values based on polynomial functions between sampled values and chooses the polynomials so that they fit smoothly together. Uses the R function `splinefun`. (i.e. makes a smooth relationship, sometimes will interpolate spikes above the data points)

Question: Under what circumstances would you consider each of these options for interpolation? 

> ANSWER: means interpolation: use with continuous data and when you don't really care about the high and low values; 
> piecewise interpolation: use when there are discrete values
> linear interpolation:
> spline interpolation: 
> would want to use linear or spline for this dataset because discharge is continuous and you would want to connect the points. When to use linear or spline: spline interpolation is based on polynomial equation and works better for physical situations, but most times in ecological situations, we don't know the polynomial, so we would likely want to use linear interpolation.
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10

Tip: Check your dataset to see if there is an NA value in the first row. You may need to add a value for that first row or trim the dataset so that the new first row corresponds to the first measurement.

```{r}
<<<<<<< HEAD
# Run a linear interpolation of the dataset to fill in gaps
USGS.flow.data.interpolated <- approx(USGS.flow.data.trimmed$datetime,
                                      USGS.flow.data.trimmed$discharge.mean, 
                                      method = "linear", n = 3630)

# Turn the interpolated dataset into a proper dataframe
USGS.flow.data.interpolated <- do.call(cbind.data.frame, USGS.flow.data.interpolated)
names(USGS.flow.data.interpolated) <- c("Date", "Discharge")
USGS.flow.data.interpolated$Date <- as.Date(USGS.flow.data.interpolated$Date, 
                                            origin = "1970/01/01")

# Create a time series object
USGS.flow.data.timeseries <- ts(USGS.flow.data.interpolated$Discharge, 
                                start = c(2009, 1) ,frequency = 12)

# Run a Seasonal Mann-Kendall test
USGS.smktest <- smk.test(USGS.flow.data.timeseries)
USGS.smktest
summary(USGS.smktest)
=======
# Run a linear interpolation of the dataset to fill in gaps; this generates a list and not a dataframe
USGS.flow.data.interpolated <- approx(USGS.flow.data.trimmed$datetime,  #approx = approximates the date and the time
                                      USGS.flow.data.trimmed$discharge.mean, 
                                      method = "linear", n = 3630) #have to tell it how many data points you want it to come up with


# Turn the interpolated dataset into a proper dataframe
USGS.flow.data.interpolated <- do.call(cbind.data.frame, USGS.flow.data.interpolated)
names(USGS.flow.data.interpolated) <- c("Date", "Discharge") #name columns because they were missing from above
USGS.flow.data.interpolated$Date <- as.Date(USGS.flow.data.interpolated$Date, 
                                            origin = "1970/01/01") #set date

# Create a time series object
USGS.flow.data.timeseries <- ts(USGS.flow.data.interpolated$Discharge, 
                                start = c(2009, 1) ,frequency = 12) #generate monthly data starting in January 2009

# Run a Seasonal Mann-Kendall test - will only take a timeseries dataset
USGS.smktest <- smk.test(USGS.flow.data.timeseries)
USGS.smktest #gives you overall trend; output shows that p-value is very small, so there is a significant trend, and a positive z-score; stronger trend the further away from zero the z-score is. 
summary(USGS.smktest) #gives us a summary of seasonal trends depending on the number of seasons we have defined (in this case it was 12); anything with one or more * means it's a significant trend
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10
```

Interpreting results of the Seasonal Mann-Kendall Test: 

  + Overall z score and p-value: test the alternative hypothesis that the true change in response variable over time is not equal to zero
  
  + Monthly z score and p-value: test the alternative hypothesis that the true change in response variable over time for a given month is not equal to zero
  
  + S: reports trend. A positive value indicates response variable increased over time, and a negative value indicates response variable decreased over time
  
Question: How would you interpret the results of the Seasonal Mann-Kendall test for this example?

<<<<<<< HEAD
> ANSWER: 
=======
> ANSWER: Over the past decade, discharge increased and then decreased each season  
>>>>>>> 734338fff605bfc3067fdd14e6ba652acccd1d10
